refactor(frontend): extract common fields from validation form
