+++ apps/frontend/components/FormValidation/FormValidation.tsx
@@ -1,21 +1,15 @@
-import { Box, TextField } from '@mui/material';
-import { DatePicker } from '@mui/x-date-pickers';
 import { DisasterDtoType, formatForm } from '@wfp-dmp/interfaces';
 import dayjs from 'dayjs';
 import { useMemo, useState } from 'react';
-import { Controller, useForm } from 'react-hook-form';
-import { useIntl } from 'react-intl';
+import { useForm } from 'react-hook-form';
 
-import { RegionFilters } from 'components/Filters/RegionFilters';
-import { DisasterSelect } from 'components/FormValidation/DisasterSelect';
+import { CommonFormFields } from './CommonFormFields';
 
 export const FormValidation = ({
   validationForm,
 }: {
   validationForm: DisasterDtoType;
 }): JSX.Element => {
-  const intl = useIntl();
-
   const formattedForm = useMemo(
     () => formatForm(validationForm),
     [validationForm],
@@ -36,107 +30,10 @@ export const FormValidation = ({
   });
 
   const [isEditMode] = useState(false);
-  const minWidth = 240;
 
   return (
     <form>
-      <Box display="flex" flexDirection="column" justifyContent={'center'}>
-        <Box display="flex" flexDirection="row">
-          <Box mr={7}>
-            <Controller
-              name="disTyp"
-              control={control}
-              render={({ field: { value, onChange } }) => (
-                <DisasterSelect
-                  disabled={!isEditMode}
-                  value={value}
-                  onChange={onChange}
-                />
-              )}
-            />
-          </Box>
-          <Controller
-            name="region"
-            control={control}
-            render={({ field: { value, onChange } }) => (
-              <RegionFilters
-                value={value}
-                onChange={onChange}
-                disableAll={!isEditMode}
-              />
-            )}
-          />
-        </Box>
-        <Box display="flex" flexDirection="row" sx={{ m: 2, mt: 5 }}>
-          <Box sx={{ mr: 6 }}>
-            <Controller
-              name="interviewer"
-              control={control}
-              render={({ field: { value, onChange } }) => (
-                <TextField
-                  disabled={!isEditMode}
-                  label={intl.formatMessage({
-                    id: 'forms_table.headers.entry_name',
-                  })}
-                  value={value}
-                  onChange={onChange}
-                  sx={{ minWidth: minWidth }}
-                />
-              )}
-            />
-          </Box>
-          <Controller
-            name="phone"
-            control={control}
-            render={({ field: { value, onChange } }) => (
-              <TextField
-                disabled={!isEditMode}
-                label={intl.formatMessage({
-                  id: 'forms_table.headers.phone',
-                })}
-                type="number"
-                value={value}
-                onChange={onChange}
-                sx={{ minWidth: minWidth }}
-              />
-            )}
-          />
-        </Box>
-        <Box display="flex" flexDirection="row" sx={{ m: 2, mt: 5 }}>
-          <Box mr={6}>
-            <Controller
-              name="reportDate"
-              control={control}
-              render={({ field: { value, onChange } }) => (
-                <DatePicker
-                  disabled={!isEditMode}
-                  label={intl.formatMessage({
-                    id: 'forms_table.headers.entry_date',
-                  })}
-                  value={value}
-                  onChange={onChange}
-                  sx={{ minWidth: minWidth }}
-                />
-              )}
-            />
-          </Box>
-          <Controller
-            name="incidentDate"
-            control={control}
-            render={({ field: { value, onChange } }) => (
-              <DatePicker
-                disabled={!isEditMode}
-                label={intl.formatMessage({
-                  id: 'forms_table.headers.dis_date',
-                })}
-                value={value}
-                onChange={onChange}
-                sx={{ minWidth: minWidth }}
-              />
-            )}
-          />
-        </Box>
-      </Box>
+      <CommonFormFields control={control} isEditMode={isEditMode} />
     </form>
   );
 };
